Make colors and battery level thresholds configurable
The colors and thresholds that dstatus uses to decorate the output
are hardcoded, requiring users to modify the source code if they
want to customize the look-and-feel.

This commit modifies dstatus so that the colors and battery level
thresholds can be configured by overriding the respective settings
in ~/.toolbox/conf/dstatus. The following configuration files and
settings are recognized.

  colors.conf
    battery_low: Color for low battery status
    battery_medium: Color for medium battery status
    battery_high: Color for high battery status
    network_name: Color used to display network names
    network_address: Color used to display network addresses
    network_disconnected: Color used for disconnected interfaces

  thresholds.conf
    battery_low: Level below which the battery is low
    battery_medium: Level below which the battery is medium